+++ Sources/SwifterSwift/UIKit/UIColorExtensions.swift
@@ -17,9 +17,12 @@ public extension UIColor {
     /// - Parameters:
     ///     - light: Color to use in light/unspecified mode.
     ///     - dark: Color to use in dark mode.
-    @available(iOS 13.0, tvOS 13.0, *)
     convenience init(light: UIColor, dark: UIColor) {
-        self.init(dynamicProvider: { $0.userInterfaceStyle == .dark ? dark : light })
+        if #available(iOS 13.0, tvOS 13.0, *) {
+            self.init(dynamicProvider: { $0.userInterfaceStyle == .dark ? dark : light })
+        } else {
+            self.init(cgColor: light.cgColor)
+        }
     }
     #endif